+++ flutile/functions.py
@@ -230,6 +230,9 @@ def ungap_indices(start, end, fasta):
             if original_index == end:
                 b = i + 1
                 break
+    # truncate to sequence length if the end index is greater than sequence length
+    else:
+      b = i + 1
     return (a, b)
 
 
@@ -317,6 +320,48 @@ def extract_ha1(subtype, *args, **kwargs):
     (start, end) = map_ha_range(start=18, end=344, subtype1=1, subtype2=subtype)
     out = _dispatch_extract(start=start, end=end, subtype=subtype, *args, **kwargs)
     smof.print_fasta(out)
+
+
+def extract_bounds(bounds, *args, **kwargs):
+    """
+    Extract a motif
+    """
+    bounds = [(min(xs), max(xs)) for xs in bounds]
+
+    # extract the interval defined by the largest and smallest indices
+    min_idx = sys.maxsize
+    max_idx = 0
+    for (a,b) in bounds:
+      min_idx = min(min_idx, a)
+      max_idx = max(max_idx, b)
+    out = _dispatch_extract(start=min_idx, end=max_idx, *args, **kwargs)
+
+    intervals = [(a - min_idx, b - min_idx + 1) for (a, b) in bounds]
+
+    pairs = []
+    for s in out:
+        defline = s.header
+        motifs = [s.seq[a:b] for (a, b) in intervals]
+        pairs.append((defline, motifs))
+
+    return pairs
+
+
+def write_bounds(named_bounds, tabular=True, *args, **kwargs):
+    bounds = [(a, b) for (name, a, b) in named_bounds]
+    pairs = extract_bounds(bounds, *args, **kwargs)
+
+    names = [name for (name, a, b) in named_bounds]
+
+    if tabular:
+        if [x for x in names if x is not None]:
+            print("\t" + "\t".join(names))
+        for (defline, motifs) in pairs:
+            print(defline + "\t" + "\t".join(motifs))
+    else:
+        smof.print_fasta(
+            [smof.FastaEntry(defline, "".join(motifs)) for (defline, motifs) in pairs]
+        )
 
 
 def gapped_indices(seq):